Обновил класс MathCast и Point3D.

+++ src/main/java/org/rendering_app/math/Point3D.java
@@ -2,9 +2,9 @@ package org.rendering_app.math;
 
 public class Point3D {
 
-    private double x;
-    private double y;
-    private double z;
+    private float x;
+    private float y;
+    private float z;
 
 
     public Point3D() {
@@ -12,7 +12,7 @@ public class Point3D {
     }
     
     
-    public Point3D(double x, double y, double z)  {
+    public Point3D(float x, float y, float z)  {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -25,35 +25,35 @@ public class Point3D {
     
     
     
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public double getZ() {
+    public float getZ() {
         return z;
     }
 
     
     
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public void setZ(double z) {
+    public void setZ(float z) {
         this.z = z;
     }
 
     
     
-    public void set(double x, double y, double z) {
+    public void set(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
